Create and use serializer mixin
